requirements.txt
alpaca-py==0.28.0
annotated-types==0.7.0
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
dill==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
frozendict==2.4.4
greenlet==3.0.3
html5lib==1.1
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
ipython==8.18.1
ipywidgets==8.1.3
jedi==0.19.1
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llvmlite==0.39.1
lxml==5.2.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
msgpack==1.0.8
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numba==0.56.4
numpy==1.23.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==3.5.3
regex==2024.5.15
requests==2.32.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
sseclient-py==1.8.0
stack-data==0.6.3
tenacity==8.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vectorbt==0.26.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
widgetsnbextension==4.0.11
yfinance==0.2.41
zipp==3.19.2
